+++ src/index.js
@@ -61,7 +61,7 @@ export { default as CommandPaletteItem } from './components/CommandPalette/Comma
 export { default as ListFilter } from './components/ListFilter/ListFilter.vue'
 export { default as Calendar } from './components/Calendar/Calendar.vue'
 export { default as NestedPopover } from './components/ListFilter/NestedPopover.vue'
-export { default as Tree } from './components/Tree.vue'
+export { default as Tree } from './components/Tree/Tree.vue'
 
 // directives
 export { default as onOutsideClickDirective } from './directives/onOutsideClick.js'
